Add USB and Improv serial support
This is a big diff that does multiple things.

First it adds USB/WebSerial support. A connection can be established through the recently redesigned main menu

It also adds "BetterSerialPorts", a small library that adds some functionality to core Serial and SerialPort web APIs to make interacting and handling them easier. Could be broken out into a standalone library if theres any desire for that.

Adds SerialConnectionList and SerialConnectionCard to show the active serial connections to the user

Adds custom hooks for interacting with SerialPort and Improv.

Adds WifiSelectionComponent as a UI for Improv provisioning

Especially SerialConnectionCard is a mess currently and needs some more cleaning up and refactoring. But overall Improv serial support is present and mostly complete.