Add ability to move controllers around
The code is still quite messy and probbably needs a round or two of refactoring and cleaning up, but it is working. In particular theres some code duplication between main.jsx and ControllerList. Maybe the append drop target need to be moved into ControllerList or the drag and drop functionality broken out into components that can be reused. This also does not have a keyboard interface yet. May look into that as part of the accessibility audit. Resolves issue #18