Add insecureOriginTemplate
insecureOriginTemplate allows for setting a template string that gets passed hostname, port, protocol, pathname, search and hash. This template string can be used to configure a more complex insuecure origin URL. This is useful for example if your application is hosted under multiple different hostnames or at multiple different pathnames. In such a situation, a single insecureOrigin url is not enough